+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+## [v0.7.2]
+
+- **Minor** Guard `renderComponent` for mounted check to avoid an un-mounted component receiving a delayed render
+- **Minor** Improve type interfaces for `RefraxComponent` / `RefraxComponentProps`
+
 # Released
 
 ## [v0.7.1](https://github.com/netarc/refrax/compare/v0.7.0...v0.7.1)